functions/authenticate_user/requirements.txt
flask
pyotp
qrcode
cryptography
pillow
pyjwt
psycopg2-binary
flask